// shared/src/models/state-machine/ISA88/ISA88StateTypeIri.ts
export enum Isa88StateTypeIri {
    Aborting = "http://www.hsu-ifa.de/ontologies/ISA-TR88#Aborting",
    Clearing = "http://www.hsu-ifa.de/ontologies/ISA-TR88#Clearing",
    Complete = "http://www.hsu-ifa.de/ontologies/ISA-TR88#Complete",
    Completing = "http://www.hsu-ifa.de/ontologies/ISA-TR88#Completing",
    Execute = "http://www.hsu-ifa.de/ontologies/ISA-TR88#Execute",
    Holding = "http://www.hsu-ifa.de/ontologies/ISA-TR88#Holding",
    Resetting = "http://www.hsu-ifa.de/ontologies/ISA-TR88#Resetting",
    Starting = "http://www.hsu-ifa.de/ontologies/ISA-TR88#Starting",
    Stopping = "http://www.hsu-ifa.de/ontologies/ISA-TR88#Stopping",
    Suspending = "http://www.hsu-ifa.de/ontologies/ISA-TR88#Suspending",
    Unholding = "http://www.hsu-ifa.de/ontologies/ISA-TR88#Unholding",
    Unsuspending = "http://www.hsu-ifa.de/ontologies/ISA-TR88#Unsuspending",

    Aborted = "http://www.hsu-ifa.de/ontologies/ISA-TR88#Aborted",
    Held = "http://www.hsu-ifa.de/ontologies/ISA-TR88#Held",
    Idle = "http://www.hsu-ifa.de/ontologies/ISA-TR88#Idle",
    Stopped = "http://www.hsu-ifa.de/ontologies/ISA-TR88#Stopped",
    Suspended = "http://www.hsu-ifa.de/ontologies/ISA-TR88#Suspended"
}