// shared/src/models/state-machine/ISA88/ISA88CommandTypeIri.ts
export enum Isa88CommandTypeIri {
    Abort = "http://www.hsu-ifa.de/ontologies/ISA-TR88#AbortCommand",
    Clear = "http://www.hsu-ifa.de/ontologies/ISA-TR88#ClearCommand",
    Hold = "http://www.hsu-ifa.de/ontologies/ISA-TR88#HoldCommand",
    Reset = "http://www.hsu-ifa.de/ontologies/ISA-TR88#ResetCommand",
    Start = "http://www.hsu-ifa.de/ontologies/ISA-TR88#StartCommand",
    Stop = "http://www.hsu-ifa.de/ontologies/ISA-TR88#StopCommand",
    Suspend = "http://www.hsu-ifa.de/ontologies/ISA-TR88#SuspendCommand",
    Unhold = "http://www.hsu-ifa.de/ontologies/ISA-TR88#UnholdCommand",
    Unsuspend = "http://www.hsu-ifa.de/ontologies/ISA-TR88#UnsuspendCommand"
}